services/pinbot/requirements.txt
redis==5.0.8
requests==2.32.3
pillow==10.4.0
python-dotenv==1.0.1
flask==3.0.3
